@@ -1,16 +1,15 @@
-import { Component, OnInit } from '@angular/core';
-import { HulafHttpServiceModule } from '@httpServices/hulafHttpServices.module';
-import { MissingPersonDto , CountryDto} from '@httpServices/hulaf-api/hulaf-api-service';
+import { Component, OnInit } from "@angular/core";
+import { HulafHttpServiceModule } from "@httpServices/hulafHttpServices.module";
+// import { MissingPersonDto , CountryDto} from '@httpServices/hulaf-api/hulaf-api-service';
 
 @Component({
-  selector: 'app-listing',
-  templateUrl: './listing.component.html',
-  styleUrls: ['./listing.component.less'],
-  providers: [HulafHttpServiceModule]
+  selector: "app-listing",
+  templateUrl: "./listing.component.html",
+  styleUrls: ["./listing.component.less"],
+  providers: [HulafHttpServiceModule],
 })
 export class ListingComponent implements OnInit {
-
-  public countries: CountryDto[];
+  /* public countries: CountryDto[];
   public missingPersons: MissingPersonDto[];
 
   constructor(hulafHttpServiceModule: HulafHttpServiceModule) {
@@ -22,9 +21,7 @@ export class ListingComponent implements OnInit {
     hulafHttpServiceModule.HulafApiService.getCountryList().subscribe(result => {
       this.countries = result;
     });
-
   }
-  ngOnInit() {
-  }
-
+  */
+  ngOnInit() {}
 }
